Fix push_state_update to only clear dirty properties on success
The `Device.push_state_update` method only pushes values of properties,
which were actually changed.  This is tracked using the `_dirty`
instance variable.

So fark, the `_dirty` flag was cleared unconditionally, before sending
messages to the AC.  This commit makes sure it is done only after
successfully sending the message.  In case of errors, the `_dirty` state
will be preserved, so a subsequent call to `push_state_update` will
re-attempt sending the same properties.