Ensure Device.update_state returns only after properties are updated
This change makes the `Device.update_state` wait for the properties
to actually be updated before returning.  The wait time is limited
by the `wait_for` parameter (which was previously present, but not
used).

This way, when `Device.update_state` returns, properties are
guaranteed to be updated.

This commit also fixes a few tests, which appear to have had some
bugs.