Ensure Device.push_state_update returns only after a getting a response from the AC
This change makes the `Device.push_state_update` wait for the a
response from the AC to be received before returning.  The wait time
is limited by the `wait_for` parameter (which was previously present,
but not used).

This way callers of `Device.push_state_update` can be sure that
their request reached the AC when the method returns.