Add the option to disable the beep when setting properties
The beep is disabled by making requests to update properties
additionally include an extra property, `Buzzer_ON_OFF` set to 1.

This setting is temporary and affects only the current update.
Subsequent requests will cause a beep unless they are sent with the
`Buzzer_ON_OFF` property set to 1 too.

This was tested with Gree Pular units.